fix(audio): harden configuration recovery
